{golem} now uses {config}
This commit bundles a series of change for relying on the {config} package inside {golem}.

There is now a file called `golem-config.yml` in the `inst/` folder, which contains the configuration for your app and that can be used for configuration in a production environment.

See the `config` Vignette for more information.