docs: move sections with relevant content for the user to the top
I. user preferences in terms of priority (more or less ...) probably
   1. Installation
   2. New project creation
   3. Every day workflow
   4. Additional resources
   5. Meta info such as about (golem version/code coverage), CoC, how to contribute etc.

II. developer preference in terms of priority (more or less ...) probably
   1. Meta info in the "About" section golem version/code coverage
   2. the rest from above

Though I'd prefer the current ordering myself it might be a wise choice to change it towards the user, but that's a matter of taste.

Fix some typos and markdown styling.